fixing a few bugs in production scripts

+++ tmp_a09m310_e_scripts/METAQ_coherent_formfac.py
@@ -20,7 +20,7 @@ parser.add_argument('-o',default=False,action='store_const',const=True,\
 parser.add_argument('-t','--t_sep',nargs='+',type=int,help='values of t_sep [default = all]')
 parser.add_argument('-d','--debug',default=False,action='store_const',const=True,\
     help='run DEBUG? [%(default)s]')
-parser.add_argument('-p','--priority',type=str,default='todo',help='put task in priority? [%(default)s]')
+parser.add_argument('-p','--priority',default=False,action='store_const',const=True,help='put task in priority? [%(default)s]')
 parser.add_argument('-v','--verbose',default=True,action='store_const',const=False,\
     help='run with verbose output? [%(default)s]')
 args = parser.parse_args()
@@ -67,6 +67,13 @@ params = {
     'M5':M5,'L5':L5,'B5':b5,'C5':c5,'MQ':mq,
     'MAX_ITER':max_iter,'RSD_TARGET':rsd_target,'Q_DELTA':delta,'RSD_TOL':rsd_tol,
     }
+if args.priority:
+    q = 'priority'
+    priority = '-p'
+else:
+    q = 'todo'
+    priority = ''
+params['PRIORITY'] = priority
 
 base_dir = '/gpfs/alpine/proj-shared/lgt100/c51/x_files/project_2/production/'+ens
 params['SCRIPT_DIR'] = '/ccs/proj/lgt100/c51/x_files/project_2/production/'+ens+'/scripts'
@@ -206,13 +213,14 @@ for c in cfgs:
                         file_time = os.stat(coherent_formfac_file_4D).st_mtime
                         if (now-file_time)/60 > time_delete:
                             print('DELETING BAD COHERENT_FF',os.path.getsize(coherent_formfac_file_4D),coherent_formfac_file_4D.split('/')[-1])
-                            shutil.move(coherent_formfac_file_4D,coherent_formfac_file_4D.replace('formfac/'+c+'/','corrupt/'))
-                            shutil.move(coherent_formfac_file,coherent_formfac_file.replace('formfac/'+c+'/','corrupt/'))
+                            shutil.move(coherent_formfac_file_4D,coherent_formfac_file_4D.replace('formfac_4D/'+c+'/','corrupt/'))
+                            if os.path.exists(coherent_formfac_file):
+                                shutil.move(coherent_formfac_file,coherent_formfac_file.replace('formfac/'+c+'/','corrupt/'))
                             #sys.exit()
                     if not os.path.exists(coherent_formfac_file) and not os.path.exists(coherent_formfac_file_4D):
                         # loop over FLAV and SPIN as all in 1 file
                         metaq  = coherent_formfac_name+'.sh'
-                        metaq_file = metaq_dir +'/'+args.priority+'/cpu/'+metaq
+                        metaq_file = metaq_dir +'/'+q+'/cpu/'+metaq
                         task_exist = False
                         task_working = False
                         if os.path.exists(metaq_file):
@@ -312,7 +320,7 @@ for c in cfgs:
                 print('    missing FLAV or SPIN seqprops, dt=',dt)
         if not have_all_seqprops:
             print('    missing FLAV or SPIN seqprops')
-            os.system('python METAQ_coherent_seqprop.py %s -v' %(c))
+            os.system('python METAQ_coherent_seqprop.py %s -v %s' %(c,priority))
         #else:
         #    print('    missing props')
     else:
